CogVM source as per VMMaker.oscog-eem.3505
Fix stupid stupid bug in jitted perform:[with:*] primitive. The code
compared the found method against the argument count of the perform:,
which is one greater than the real argument count (since the selector
is dropped). This bug is two years old. I am a bear of very little brain.