CogVM source as per VMMaker.oscog-eem.3512
Introduce eeInstantiatePoint for Spur to use Point's pre-defined class index.
Improve Slang inlining of constant expressions so that YIndex + 1 is inlined.

Add a special case for pinObject:, and don't reuse rcvr to stablilize
primitiveBeDisplay's accessor depth calculation.

Use the same convention for unscannedEphemerons as for savedFirstFieldsSpace.

SmartSyntaxPlugins
It is high time the generated type for ByteArray in SmartSyntaxPlugins is
unsigned char * instead of the ambiguous char *. This is motivated by type
warnings in the VectorEnginePlugin, and by the fact that ByteArray is an
UnsignedIntegerArray subclass.

Slang:
Avoid generating failure guards for floatObjectOf:; it can't fail.
Add support for squared for the VectorEnginePlugin.