macOS: fix the regression in start-up which delivers a DnD event for a file argument on launch. This is needed only in the Terf/Virtend VM.
Refactor the declaration of the OSSemaphore type so that the "single-
threaded" mac VM (which now has a GUI thread in it) can access the
Smalltalk thread from the GUI thread to print a stack trace.

Fix and update the relevant VMMaker and Reader image scripts.

In the reader image, the threaded-FFI VM now evaluates 3+4 in
parallel with the Squeak UI being usable.