CogVM source as per VMMaker.oscog-eem.3539
The Spur64 Windows 11 arm64 crash is caused by the hash of Array being set to zero
(it should be the known class index 51). A fix is to access the marked bit via byte
access, not 32-bit halfword access. This commit along with VMMaker.oscog-eem.3537
give a chance of locating the bad code caused by the bug.

Don't translate flag (it's defined as an empty macro in sqMemoryAccess.h anyway).
This helps isMarked: and numSlotsOf: to be inlined. Mark setIsMarkedOf:to: as inlined.