CogVM source as per VMMaker.oscog-eem.3545
On ARM64 if homogenous float aggregates can be passed in registers then their
fields are spread across the least-significant halves of the outgoing double-
precision float argument registers.

Thanks to Leonard Geier and Tom Beckmann for this fix.

BTW, generating the RiscV64FFIPlugin also as this is written as a subclass
of ThreadedARM64FFIPlugin. This may or may not be correct. Ken Dickey can
you check?

P.S. we need the FFI test suite extended with tests for homogenous float
aggregates.