CogVM source as per Cog-eem.473
Refactor the processor simulator alien tests so that (almost all of) the tests are
implemented in CogProcessorSimulatorTest, which consequenty provides much clearer
definition of required semantics. Add full description to platforms/Cross/plugins/
ProcessorSimulatorPlugin.h.  Fix some issues with the plugins not terminating
correctly when stepping from executable to non-executable memory. Fix a bad bug
in the COG hack of bochs (cpu_step) which always committed the rip as the next_rip
even if it was out-of-bounds. Nuke some extra nonsense in the arm processor plugin
iOS makefiles (cuz access to assert from plugins has been fixed for a while).

remove dependency on http://ss3.gemtalksystems.com when building a VMMaker image
now that the freetype stuff is cloned on squeaksource.