macOS threaded GUI:
Complement a117332f22daa0f64413a235696c2a3ec7b264fc (using a GUI thread
to simplify threading on macOS). Use of the threaded mac VM on Cuis
revealed major slowdowns. The main issue was the Mac being the only
platform not to allow deferring display updates. Because Spur VMs pin the
display bitmap cannotDeferDisplayUpdates does not need to be set on mac,
eliminating the slowdown. But when cannotDeferDisplayUpdates is set the
slowdown is due to squeakUIFlushPrimaryDeferNMilliseconds. This was
33.3ms. Given modern machines this has been reduced to 1ms.

One update path to the display used runBlockOnMainThread: to invoke
setNeedsDisplayInRect:. Since this only happens on the GUI thread it can
be invoked directly.

runBlockAsyncOnMainThread: can use a service modifier to be more prompt.