CogVM source as per VMMaker.oscog-eem.3568
Fix Slang bug with value expansions that have accidental type mismatches between
block argument actuals and other varianles (affects the Newspeak version of
unlinkIfLinkedSend:pc:to: and targetMethod, which is used both as a sqInt and as
a CogMethod *.  Consequently regenerate sources based on VMMaker.oscog-mt.3568
which is a discarded branch of a hack fix.

Ensure that the bochs processor simulators handle panics in resetCPU without
crashing the VM. Nuke some old configuration files. Update a README