CogVM source as per VMMaker.oscog-eem.3583
Revise optional/essential pragma marking of methods so that a method
can both be essential in one configuration and included optionally in
another. In this case primitiveAllMethodsCompiledToMachineCode is both
essential to the Sista VM and needed for VMInvestigations. Now that
Slang block support is adequate, use methodsDo: for
methodsCompiledToMachineCodeInto: instead of hand-coding the enumeration.

So introduce the essential: pragma, offloading option: for marking
methods that must be included in certain configurations, allowing
it to mark truly optional methods.

Reorganize Lowcode methods so that Lowcode is only supported in Spur
(we don't generate a V3 Lowcode VM and no one has asked for it).

primitiveBytecodeSetsAvailable should answer one or two values. The
VM supports EncoderForV3PlusClosures, which is a strict superset of
EncoderForV3. Therefore the VM should *not* answer EncoderForV3 ever.
Document that Interpreter supports EncoderForV3PlusClosures by
giving it an implementation, albeit with a single set.