CogVM source as per VMMaker.oscog-eem.3595
StackInterpreter:
When invoking setInterruptPending: make sure to set interruptPending before
calling forceInterruptCheck so that routines on the interruptCheckChain can
test for interruptPending.

Fix the AndreasSystemProfiler in the StackInterpreter. The key bug is that
the nextProfileTick must not be set to zero until the profileSemaphore has
been signalled, otherwise if routines like checkProfileTickPostPrimitive:
take a sample and do forceInterruptCheck, when the interrupt check routine
is called it will do nothing (because nextProfileTick will be zero). So
checkProfileTickIfSwitched: et al only take a sample if profileProcess is nil.

Another bug was that if the number of ticks chosen for the profile (i.e.
ticks of the high resilution clock) are few then the first sample was the
valueNoContextSwitch in the ensure: in AndreasSystemProfiler>>spyOn:. So
add a flag which is set in the no-context-switch value primitives which
defers taking a sample in checkProfileTickPostPrimitive:. This ensures
that the first sample taken will be in the spied-on block, not in the
AndreasSystemProfiler machinery.

Another bug was that my rewrite failed to nil profileProcess & profileMethod,
which could lead to memory leaks. Here save a little by nilling profileProcess
in checkProfileTickIfSwitched:, but only nilling profileMethod in the tracing
routines if profileProcess is nil. (profileProcess isNil implies profileMethod
isNil).

For speed and space don't use objectMemory nilObject as the nil marker in
profileSemaphore, profileProcess & profileMethod. It's easier to simply use nil.

Don't inline #followForwarded: in followForwardingPointersInProfileState and
some other non-time-critical places.

Add class vars for the no check value primitive numbers so that profiling
can avoid profiling them.

Fix a refactoring slip in checkHeapIntegrity:classIndicesShouldBeValid: