CogVM source as per VMMaker.oscog-eem.3603
Fix the AndreasSystemProfiler. This profiler has two variables,
profileProcess and profileMethod, that are sampled on a profile
tick tested after each primitive invocation and in each event
check. The old JIT code only sampled primitiveMethod on failure,
which is exactly the wrong thing to do. Make sure to nil
profileProcess and profileMethod in primitiveProfileStart
(which sets the next tick expiry).

Use nil for the unset values of profileProcess, profileMethod,
and profileSemaphore. Use the Slang support to update these via
profileStateDo[Updating]:

In the Cog VM revise flushing so that only contexts that have
profiled primitives (and hence those primitives have failed)
have pcs mapped to bytecode pcs.  Consequently rewrite
primitivePropertyFlags:numArgs: into primitivePropertyFlags:primIndex:.

When profiling, for simplicity, call all normal primitives via the
sideways call, leaving it to cePrimReturnEnterCogCodeProfiling to
take the same post primitive. Only fgo the extra mile to take a
sample post primitive in the code for C primitives compiled to
run on the Smaltalk stack.