CogVM source as per VMMaker.oscog-eem.3610
Perfect Andreas' profiler by delivering primitive samples immediately.
The old code took the sample after invoking a primtiive but delivered
it at the next interrupt check (frame building send or backward jump).
Imagine a method answering the result of a send that invokes a primitive.
The sample will be taken, but not delivered immediately. Instead the
sender will return and a subsequent send will check for events, causing
the sample to be delivered. Hence the primitive will not be attributed
to the actual sender.

The change is simply to do the semaphore signal in the primitive
sampling routine, activating the profile process immediately, and
leaving the sender suspended immediately after the primitive invoking send.

There are consequences for on-stack primitive invocation
(see compileOnStackExternalPrimitive:flags:).

Have primitiveProfileSemaphore zero nextProfileTick when installing
nil as the semaphore. nextProfileTick can then be used as a guard of
profileSemaphore avoiding invoking synchronousSignal with nil.

Refactor checkProfileTickIfSwitched: to checkProfileTickIfSwitched
for neater code in checkForEventsMayContextSwitch:.

Now that the StackInterpreter does logging CoInterpreter no longer
needs to override synchronousSignal:, and we only need one trace
flag for signals.

CogARMv8Instruction: fix a bad bug with using MOVN for MoveCq:R:. The old code could produce nonsense instructions because it didn't mask negative values to the 16-bit field width.