CogVM source as per VMMaker.oscog-eem.3635
Fix an assert and the comment in updateStateOfSpouseContextForFrame:WithSP:

Move the branch offset calculators up to Cogit so CogitForDeferredDisassembly
can work but inherit only from Cogit.

Get deferred disassembly (the disassembly of data obtained via
primitiveCogitMethod) working on ARMv8.  Export static symbolic names
for deferred disassembly via primitiveCogStaticSymbols.

Slang:
Fix a subtle and long-standing bug in type inferrence that caused occasional
instability.

When typing returns in inferReturnTypeFromReturnsIn: the search would abort
if it found an as-yet-untyped return expression. This is the correct thing
to do until there is no more progress to be made, as gradually most, if not
all, of those as-yet-untyped returns should become typed. At this point a
return type must be chosen from the typed return expressions so that
methods typing variables based on method return types can do so.

An example result of this bug was to type methodOrSelectorIndex as int, not
that other methods it invoked were not yet typed by the end of the phase.

The fix is to replace inferReturnTypeIn: and inferReturnTypeFromReturnsIn:
by inferReturnTypeIn:finalPass: and inferReturnTypeFromReturnsIn:finalPass:
which do not abort if finalPass is true, and for inferTypesForImplicitlyTypedVariablesAndMethods
to ensure that a final pass has been made with finalPass true.

When inferring implicitly typed variables (inferTypesForImplicitlyTypedVariablesIn:),
type untyped literal block arguments where possible.

Consequently provide Cogit>>trampolineRangesDo: which then allows the tokenization tests to be moved to CogDisassemblerDecorator.

methodAfter: should use asUnsignedInteger

CogMethodConstants does not need to be a shared pool of NewCoObjectMemory,
Spur[32|64]BitCoMemoryManager, or StackInterpreter. CoInterpreter can avoid
using CMBlock and CMMethod, except in one case (getCogMethodCount).

UpdateSqueakTrunkImage.st: try and set the transcript'c character limit
so that e.g. update crashes can be comprehended.