CogVM source as per VMMaker.oscog-eem.3660
Fix the imbecillic regression in VMMaker.oscog-eem.3637 which should
have touched just the return type of stackPointerIndexForFrame:WithSP:
but instead broke it for interpreter frames.
We now have a working Windows ARMv8 VM!!

Refactor checkStackIntegrity so we only need
StackInterpreter>>checkStackIntegrity.

Simplify cmHomeMethod. No need to test cmIsFullBlock.

Eliminate the assert in cogMethodContaining:. Now we use it in
click step printing it is used in many more circumstances.