CogVM source as per VMMaker.oscog-eem.3667
Fix Nico Papagna's BitBlt clipping bug which was due to use on int variables
on 64-bits so that negative values were truncated to extremely high positive
values, defeating the clipping logic.

As a result redefine the coordinate parameters of the surface functions to
take sqInts.

Change the signature of sqResolverGetAddressInfoResultSize
to eliminate a warning.

Slang: avoid adding struct type aliases for abstract classes,
eliminating some warnings in the ThreadeedFFI plugins.