CogVM source as per VMMaker.oscog-eem.3681
Fix the AndreasSystemProfiler misattribution bug, which is caused by the tally
being sampled. The fix requires that runProfileProcess zero nextProfileTick
while profiling.  The fix also requires that primitiveProfileStart does not
nil the samples (profileProcess and profileMethod).

Since checkProfileTickPostPrimitive: is always invoked with newMethod,
remove the argument and access newMethod directly.