CogVM source as per VMMaker.oscog-eem.3697
Interpreter: eliminate a warning in primitiveHighBit on 64-bit Windows.

Slang: do a better job translating printf format strings, avoiding replacing
format specifers with the PRI?SQINT for non-sqInt parameters. This reduces
warnings on windows 64bit where suzeof(long) = 4 and long ~= sqInt.

Change the initialization of especial CogStackPage to get BytesPerWord
initialized correctly when a CCodeGenerator is created in
buildCodeGeneratorFor[Cogit|Interpreter]:includeAPIMethods:initializeClasses:.
This is necessary to get kernelReturnTypes correct. Now the return type of
long32At: is int, and consequently so is rawHashBitsOf: