CogVM source as per VMMaker.oscog-eem.3710
Add a machine-code primitiveSlotAtPut for 32-bit Spur (only
deals with SmallInteger values assigned to 64-bit slots).

Make the interpreter versions of primitiveSlotAtPut accept
64-bit values assigned to 32-bit slots.

Implement 64-bit DoubleWordArray access for machine-code
primitiveSlotAt on 32-bit Spur.

Fix awful bug in machine-code primitiveAtPut on 32-bit Spur
with DoubleWordArrays which forgot to zero the top 4 bytes
of a slot if assigned a SmallInteger value.