CogVM source as per VMMaker.oscog-eem.3713
Fix bad regression in VMMaker.oscog-eem.3709. The fix for
double word at put not zeroing MSW on 32-bit Spur was wrong
in CogObjectRepresentationFor32BitSpur>>genPrimitiveAtPut.
Do not use the unavailable-on-IA32 Extra0Reg in
CogObjectRepresentationFor32BitSpur>>genPrimitiveSlotAtPut.

Use f:printf: instead of print: in printPrimLogEntryAt:hasParameter:
to get \n to p[rint as a newline in the simulator.