CogVM source as per VMMaker.oscog-eem.3716
Cogit:
rename genAlloc generqtors to genInstantiate
(since they create initialized instances).
For Spur add a mayFail parameter and pass false as often
as possible. Hence eliminate scavengeThresholdAddress in Spur
as it is now unused (scavengeThreshold is a constant in Spur).
Since the instantiators save a register by comparing against
the scavengeThreshold minus an offset, fix the asserts in
the SpurCoMemoryManager ceScheduleScavenge entry points.

Get rid of the isBackwardBranchFixup rename for isTargetOfBackwardBranch